+++ server/doppelkopf/datatypes.py
