+++ server/doppelkopf/errors.py
